Move the Host extractor to axum-extra
