+++ axum-extra/src/extract/mod.rs
@@ -1,7 +1,9 @@
 //! Additional extractors.
 
 mod cached;
+mod host;
 mod optional_path;
+pub mod rejection;
 mod with_rejection;
 
 #[cfg(feature = "form")]
@@ -19,7 +21,9 @@ mod query;
 #[cfg(feature = "multipart")]
 pub mod multipart;
 
-pub use self::{cached::Cached, optional_path::OptionalPath, with_rejection::WithRejection};
+pub use self::{
+    cached::Cached, host::Host, optional_path::OptionalPath, with_rejection::WithRejection,
+};
 
 #[cfg(feature = "cookie")]
 pub use self::cookie::CookieJar;
