+++ axum/src/extract/mod.rs
@@ -10,7 +10,6 @@ pub mod rejection;
 #[cfg(feature = "ws")]
 pub mod ws;
 
-mod host;
 pub(crate) mod nested_path;
 mod raw_form;
 mod raw_query;
@@ -26,7 +25,6 @@ pub use axum_macros::{FromRef, FromRequest, FromRequestParts};
 #[doc(inline)]
 #[allow(deprecated)]
 pub use self::{
-    host::Host,
     nested_path::NestedPath,
     path::{Path, RawPathParams},
     raw_form::RawForm,
