+++ axum/src/extract/rejection.rs
@@ -67,14 +67,6 @@ define_rejection! {
 
 define_rejection! {
     #[status = BAD_REQUEST]
-    #[body = "No host found in request"]
-    /// Rejection type used if the [`Host`](super::Host) extractor is unable to
-    /// resolve a host.
-    pub struct FailedToResolveHost;
-}
-
-define_rejection! {
-    #[status = BAD_REQUEST]
     #[body = "Failed to deserialize form"]
     /// Rejection type used if the [`Form`](super::Form) extractor is unable to
     /// deserialize the form into the target type.
@@ -175,16 +167,6 @@ composite_rejection! {
     pub enum RawPathParamsRejection {
         InvalidUtf8InPathParam,
         MissingPathParams,
-    }
-}
-
-composite_rejection! {
-    /// Rejection used for [`Host`](super::Host).
-    ///
-    /// Contains one variant for each way the [`Host`](super::Host) extractor
-    /// can fail.
-    pub enum HostRejection {
-        FailedToResolveHost,
     }
 }
